move isomorphic migration code to separate file
